+++ Sources/ApiKit/ApiEnvironment.swift
@@ -8,20 +8,19 @@
 
 import Foundation
 
-/**
- This protocol can be implemented to define API environments
- or specific API versions.
-
- An ``ApiEnvironment`` must define a root ``url`` to which a
- route ``ApiRoute/path`` can be applied.
- 
- You can use enums to define environments and use associated
- values to provide environment-specific parameters.
-
- Both ``ApiEnvironment`` and ``ApiRoute`` can define headers
- and query parameters. An environment can use this to define
- global data, while a route defines route-specific data.
- */
+/// This protocol can be used to define API environments, or
+/// specific API versions.
+///
+/// An ``ApiEnvironment`` must define a ``url``, to which an
+/// environment-relative ``ApiRoute/path`` can be appended.
+///
+/// You can use enums to define environments, and associated
+/// values to provide environment-specific parameters.
+///
+/// Both ``ApiEnvironment`` and ``ApiRoute`` can specify any
+/// headers and query parameters they need. Environments can
+/// define global headers and query parameters, while routes
+/// can define route-specific ones.
 public protocol ApiEnvironment: ApiRequestData {
 
     /// The base URL of the environment.