+++ Sources/ApiKit/ApiError.swift
@@ -8,19 +8,20 @@
 
 import Foundation
 
-/**
- This enum defines api-specific errors that can occur when a
- client communicates with an external API.
- */
+/// This enum defines api-specific errors that can be thrown
+/// when an ``ApiClient`` communicates with any external API.
 public enum ApiError: Error, Equatable, LocalizedError {
     
     /// This error should be thrown when an ``ApiEnvironment``
     /// has a url that can't be used to generate a `URL`.
     case invalidEnvironmentUrl(String)
     
-    /// This error should be thrown when a URL request fails
-    /// because of an invalid response status code.
+    @available(*, deprecated, renamed: "unsuccessfulHttpStatusCode")
     case invalidResponseStatusCode(Int, URLRequest, URLResponse, Data)
+
+    /// This error should be thrown when a URL request fails
+    /// due to an invalid status code (outside of 100-599).
+    case invalidHttpStatusCode(Int, URLRequest, URLResponse, Data)
 
     /// This error should be thrown when a `URLRequest` will
     /// fail to be created due to invalid `URLComponents`.
@@ -29,13 +30,19 @@ public enum ApiError: Error, Equatable, LocalizedError {
     /// This error should be thrown when a `URLRequest` will
     /// fail to be created due to an invalid `URL`.
     case failedToCreateComponentsFromUrl(URL)
+
+    /// This error should be thrown when a URL request fails
+    /// due to an unsuccessful status code (100-199, as well
+    /// as 300-599).
+    case unsuccessfulHttpStatusCode(Int, URLRequest, URLResponse, Data)
 }
 
 public extension ApiError {
-    
+
+    @available(*, deprecated, message: "This will be removed in 1.0. Switch over the error and use `unsuccessfulHttpStatusCode` instead.")
     var isInvalidResponseStatusCode: Bool {
         switch self {
-        case .invalidResponseStatusCode: true
+        case .unsuccessfulHttpStatusCode: true
         default: false
         }
     }