+++ Sources/ApiKit/Extensions/String+UrlEncode.swift
@@ -12,24 +12,20 @@ import Foundation
 
 extension String {
 
-    /**
-     Encode the string to work with `x-www-form-urlencoded`.
-
-     This will first call `urlEncoded()`, then replace every
-     `+` with `%2B`.
-     */
+    /// Encode the string to work for `x-www-form-urlencoded`.
+    ///
+    /// This will first call `urlEncoded()` and replace each
+    /// `+` with `%2B`.
     func formEncoded() -> String? {
         self.urlEncoded()?
             .replacingOccurrences(of: "+", with: "%2B")
     }
     
-    /**
-     Encode the string to work with quary parameters.
-     
-     This will first call `addingPercentEncoding`, using the
-     `.urlPathAllowed` character set, then replace every `&`
-     with `%26`.
-     */
+    /// Encode the string for quary parameters.
+    ///
+    /// This will first call `addingPercentEncoding` using a
+    /// `.urlPathAllowed` character set then replace each `&`
+    /// with `%26`.
     func urlEncoded() -> String? {
         self.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)?
             .replacingOccurrences(of: "&", with: "%26")