+++ Sources/ApiKit/Integrations/Yelp/Yelp.swift
@@ -8,11 +8,8 @@
 
 import Foundation
 
-/**
- This namespace contains a Yelp-specific set of environments,
- routes and data types.
- 
- You can setup a dev account at `https://yelp.com/developers`
- to start using these types with ApiKit.
- */
+/// This namespace contains a Yelp-specific types.
+///
+/// You can read more on how to set up a test API account at
+/// `https://yelp.com/developers`.
 public struct Yelp {}